修复了warning（虚基类析构函数，size_t，case break）；使用.at()访问动态数组确认了越界问题

+++ syntax.tab.cpp
@@ -2228,7 +2228,7 @@ yyreturnlab:
 
 
 int yylex() {
-    if (tokenIndex >= tokens.size()) return 0;
+    if (tokenIndex >= static_cast<int>(tokens.size())) return 0;
     Token* current = &tokens[tokenIndex];
     yylval.token = current;
